# contracts/compat/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Vondel Private Compatibility Service API
  version: 1.0.0
  description: >
    Private, capability-scoped API consumed exclusively by enrolled Vondel
    compatibility applications (vondel-audiobookshelf, vondel-jellyfin). It is
    served below /api/internal/compat/v1 and is never part of the public route
    surface: the edge gateway must not forward /api/internal/** from public
    ingress. Every operation authenticates the calling application, enforces
    its granted capability, and — when subject-scoped — resolves the acting
    subject exclusively from the signed, audience-bound subject token. An
    application can never name the subject it acts for. Mutations require
    idempotency keys; lists use signed cursors; media and artwork resolve to
    short-lived, audience-bound, same-origin delivery grants.
servers:
  - url: /api/internal/compat/v1
security:
  - serviceBearer: []

# Reusable blocks -----------------------------------------------------------
...
paths:
...
  /identity/switch:
    post:
      operationId: switchProfile
      summary: Switch to another authorized profile, minting a new subject token.
      description: >
        The target profile must be authorized for the subject's device grant.
        A PIN-protected profile requires its PIN; an unprotected profile
        switches without one. The previous subject token is not revoked by the
        switch itself.
      x-vondel-capability: identity
      x-vondel-subject: true
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/SubjectToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchProfileRequest"
      responses:
        <<: *stdErrors
        "200":
          description: New subject token for the switched profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
...
components:
  securitySchemes:
    serviceBearer:
      type: http
      scheme: bearer
      description: Short-lived renewable service credential issued at enrollment.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >
        Caller-chosen key making the mutation safely retryable. A replay with
        the same key and payload returns the stored response; the same key
        with a different payload is a conflict.
      schema:
        type: string
        minLength: 1
        maxLength: 200
    SubjectToken:
      name: X-Vondel-Subject-Token
      in: header
      required: true
      description: >
        Signed, audience-bound subject token minted by login or profile
        switch. It is the only way a subject is named; the server revalidates
        the subject on every request.
      schema:
        type: string
...
  schemas:
...
    DeviceClaim:
      type: object
      required: [id]
      properties:
        id:
          type: string
        name:
          type: string
        platform:
          type: string
        ip_address:
          type: string
    Subject:
      type: object
      required:
        [
          account_id,
          profile_id,
          organization_id,
          membership_id,
          policy_revision,
          security_revision,
          credential_revision,
          auth_method,
          device,
        ]
      properties:
        account_id:
          type: integer
        profile_id:
          type: string
        organization_id:
          type: string
        membership_id:
          type: string
        policy_revision:
          type: integer
          format: int64
        security_revision:
          type: integer
          format: int64
        credential_revision:
          type: integer
          format: int64
        auth_method:
          type: string
        device:
          $ref: "#/components/schemas/DeviceClaim"
...
    LoginResult:
      type: object
      required: [subject_token, expires_at, subject]
      properties:
        subject_token:
          type: string
        expires_at:
          type: string
          format: date-time
        subject:
          $ref: "#/components/schemas/Subject"
...
    SwitchProfileRequest:
      type: object
      required: [profile_id]
      properties:
        profile_id:
          type: string
        pin:
          type: string
...